% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasNN.R
\name{lasNN}
\alias{lasNN}
\title{Merge neighbors variables by a function}
\usage{
lasNN(las, las_from, k = 1, func = "mean", var = "Intensity",
  name = "V", desc = "Description for V")
}
\arguments{
\item{las}{an object of class LAS to be the reference}

\item{las_from}{an object of class LAS from which neighbors will be aggregated}

\item{k}{interger. The number of k-nearest neighbors.}

\item{func}{character.A function name to aggregate the neighbors values, accepted values
("mean", "median", "q1", "q3", "var", "sd", "max", "min"), default "mean"}

\item{var}{string. The name of a point variable to aggregate data from.}

\item{name}{string. The name variable to store the result.}

\item{name}{string. Description for the variable.}
}
\value{
A numeric vector with the neighbors aggregated value for a given variable
}
\description{
Merge neighbors variables by a function
}
