% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasRANN-package.r
\name{lasRANN-package}
\alias{lasRANN-package}
\alias{lasRANN}
\title{Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library}
\description{
Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library
}
