% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RANN2-package.r
\name{RANN2-package}
\alias{RANN2-package}
\alias{RANN2}
\title{Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library}
\description{
Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library
}
\seealso{
\code{\link{nn2}}
}
