% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RANN2-package.r
\docType{package}
\name{RANN2-package}
\alias{RANN2-package}
\alias{_PACKAGE}
\alias{RANN2}
\title{RANN2: Fast Nearest Neighbour Search ('Rcpp' Wrapper of 'Arya' and Mount's 'libANN')}
\description{
Finds the k nearest neighbours for every point in a given dataset
    in O(N log N) time using 'Arya' and Mount's 'ANN' library (v1.1.3). This wrapper
    uses 'Rcpp' to simplify access compared with the original 'RANN' package.
}
\seealso{
\code{\link{nn2}}
}
\author{
\strong{Maintainer}: Gregory Jefferis \email{jefferis@gmail.com} (\href{https://orcid.org/0000-0002-0587-9355}{ORCID})

Authors:
\itemize{
  \item Sunil Arya (\href{https://orcid.org/0000-0003-0939-4192}{ORCID}) [copyright holder]
  \item David Mount (\href{https://orcid.org/0000-0002-3290-8932}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Caio Hamamura (\href{https://orcid.org/0000-0001-6149-5885}{ORCID}) [contributor]
  \item Jurian Baas (\href{https://orcid.org/0000-0001-8689-8824}{ORCID}) [contributor]
  \item University of Maryland (ANN library is copyright University of Maryland and Sunil Arya and
    David Mount. See file COPYRIGHT for details) [copyright holder]
}

}
\keyword{internal}
