% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RANN2-package}
\alias{RANN2}
\alias{RANN2-package}
\title{Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library}
\description{
Wrapper for Arya and Mount's Approximate Nearest Neighbours (ANN) C++ library
}
\seealso{
\code{\link{nn2}}
}

